backend/requirements-pytorch.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.2.0
numpy>=1.26.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
reportlab>=4.0.0
pandas>=2.0.0
google-generativeai>=0.3.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
email-validator>=2.0.0
